requirements.txt
streamlit==1.54.0
numpy==2.3.3
opencv-python-headless
pillow
tensorflow==2.21.0